app/requirements.txt
python-jobspy
pyinstaller
